Add support for Blackthorne pack file format
Blackthorne uses a very similar pack file format to The Lost Vikings.
Add a flag to the pack file structure which indicates which format it
is. The pack_tool and sprite_view tools can now be used for both games.

Also added support for loading the palette from a chunk in the
sprite_view tool.